# src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Analisale Documentation",
    "description": "This is the documentation for the Analisale API (in developmeent)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:3000/users",
      "description": "User API"
    },
    {
      "url": "https://localhost:3000/products",
      "description": "Product API"
    },
    {
      "url": "https://localhost:3000/sales",
      "description": "Sale API"
    },
    {
      "url": "https://localhost:3000/charges",
      "description": "Charge API"
    }
  ],
...
  "paths": {
...
    "/charges": {
      "get": {
        "tags": ["Charge"],
        "summary": "Get all charges",
        "description": "This route will list all charges created",
        "operationId": "getAllCharges",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Charge"
                }
              }
            }
          },
          "404": {
            "description": "Charge not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "post": {
        "tags": ["Charge"],
        "summary": "Add a new charge",
        "description": "This route will create a new charge",
        "operationId": "createCharge",
        "requestBody": {
          "description": "Create a new product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Charge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Charge"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/charges/{chargeId}": {
      "get": {
        "tags": ["Charge"],
        "summary": "Find charge by ID",
        "description": "Returns a single charge by its ID",
        "operationId": "getChargeById",
        "parameters": [
          {
            "name": "chargeId",
            "in": "path",
            "description": "ID of charge to return",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Charge"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["Charge"],
        "summary": "Deletes charge by ID",
        "description": "This route will delete a charge by its id",
        "operationId": "deleteCharge",
        "parameters": [
          {
            "name": "chargeId",
            "in": "path",
            "description": "Charge id to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Charge"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID value"
          }
        }
      },
      "put": {
        "tags": ["Charge"],
        "summary": "Update charge by ID",
        "description": "",
        "operationId": "updateCharge",
        "parameters": [
          {
            "name": "chargeId",
            "in": "path",
            "description": "ID of charge to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Charge"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "Produto A"
          },
          "type": {
            "type": "integer",
            "example": 1
          },
          "amount": {
            "type": "integer",
            "example": 7
          },
          "salePurchase": {
            "type": "integer",
            "example": 20.99
          }
        }
      },
      "Charge": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Tax"
          },
          "amount": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    },
...
  }
}